TSCH: rework .h system, now following the same structure as rpl-lite
