+++ examples/6tisch/sixtop/node-sixtop.c
@@ -42,8 +42,6 @@
 #include "sys/log.h"
 #include "net/ipv6/uip-ds6-route.h"
 #include "net/mac/tsch/tsch.h"
-#include "net/mac/tsch/tsch-log.h"
-#include "net/mac/tsch/tsch-schedule.h"
 #include "net/mac/tsch/sixtop/sixtop.h"
 #include "net/routing/routing.h"
