+++ examples/6tisch/sixtop/sf-simple.c
@@ -40,7 +40,7 @@
 #include "contiki-lib.h"
 
 #include "lib/assert.h"
-#include "net/mac/tsch/tsch-schedule.h"
+#include "net/mac/tsch/tsch.h"
 #include "net/mac/tsch/sixtop/sixtop.h"
 #include "net/mac/tsch/sixtop/sixp.h"
 #include "net/mac/tsch/sixtop/sixp-nbr.h"
