+++ os/net/mac/tsch/tsch-log.c
@@ -48,12 +48,6 @@
 #include "contiki.h"
 #include <stdio.h>
 #include "net/mac/tsch/tsch.h"
-#include "net/mac/tsch/tsch-queue.h"
-#include "net/mac/tsch/tsch-private.h"
-#include "net/mac/tsch/tsch-log.h"
-#include "net/mac/tsch/tsch-packet.h"
-#include "net/mac/tsch/tsch-schedule.h"
-#include "net/mac/tsch/tsch-slot-operation.h"
 #include "lib/ringbufindex.h"
 #include "sys/log.h"
