+++ os/net/mac/tsch/tsch-log.h
@@ -42,7 +42,6 @@
 
 #include "contiki.h"
 #include "sys/rtimer.h"
-#include "net/mac/tsch/tsch-private.h"
 
 /******** Configuration *******/
