+++ os/net/mac/tsch/tsch-queue.c
@@ -53,11 +53,6 @@
 #include "lib/random.h"
 #include "net/queuebuf.h"
 #include "net/mac/tsch/tsch.h"
-#include "net/mac/tsch/tsch-private.h"
-#include "net/mac/tsch/tsch-queue.h"
-#include "net/mac/tsch/tsch-schedule.h"
-#include "net/mac/tsch/tsch-slot-operation.h"
-#include "net/mac/tsch/tsch-log.h"
 #include <string.h>
 
 /* Log configuration */
