+++ os/net/mac/tsch/tsch-queue.h
@@ -43,122 +43,7 @@
 #include "contiki.h"
 #include "lib/ringbufindex.h"
 #include "net/linkaddr.h"
-#include "net/mac/tsch/tsch-schedule.h"
 #include "net/mac/mac.h"
-
-/******** Configuration *******/
-
-/* The maximum number of outgoing packets towards each neighbor
- * Must be power of two to enable atomic ringbuf operations.
- * Note: the total number of outgoing packets in the system (for
- * all neighbors) is defined via QUEUEBUF_CONF_NUM */
-#ifdef TSCH_QUEUE_CONF_NUM_PER_NEIGHBOR
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR TSCH_QUEUE_CONF_NUM_PER_NEIGHBOR
-#else
-/* By default, round QUEUEBUF_CONF_NUM to next power of two
- * (in the range [4;256]) */
-#if QUEUEBUF_CONF_NUM <= 4
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 4
-#elif QUEUEBUF_CONF_NUM <= 8
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 8
-#elif QUEUEBUF_CONF_NUM <= 16
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 16
-#elif QUEUEBUF_CONF_NUM <= 32
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 32
-#elif QUEUEBUF_CONF_NUM <= 64
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 64
-#elif QUEUEBUF_CONF_NUM <= 128
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 128
-#else
-#define TSCH_QUEUE_NUM_PER_NEIGHBOR 256
-#endif
-#endif
-
-/* The number of neighbor queues. There are two queues allocated at all times:
- * one for EBs, one for broadcasts. Other queues are for unicast to neighbors */
-#ifdef TSCH_QUEUE_CONF_MAX_NEIGHBOR_QUEUES
-#define TSCH_QUEUE_MAX_NEIGHBOR_QUEUES TSCH_QUEUE_CONF_MAX_NEIGHBOR_QUEUES
-#else
-#define TSCH_QUEUE_MAX_NEIGHBOR_QUEUES ((NBR_TABLE_CONF_MAX_NEIGHBORS) + 2)
-#endif
-
-/* TSCH CSMA-CA parameters, see IEEE 802.15.4e-2012 */
-/* Min backoff exponent */
-#ifdef TSCH_CONF_MAC_MIN_BE
-#define TSCH_MAC_MIN_BE TSCH_CONF_MAC_MIN_BE
-#else
-#define TSCH_MAC_MIN_BE 1
-#endif
-/* Max backoff exponent */
-#ifdef TSCH_CONF_MAC_MAX_BE
-#define TSCH_MAC_MAX_BE TSCH_CONF_MAC_MAX_BE
-#else
-#define TSCH_MAC_MAX_BE 5
-#endif
-/* Max number of re-transmissions */
-#ifdef TSCH_CONF_MAC_MAX_FRAME_RETRIES
-#define TSCH_MAC_MAX_FRAME_RETRIES TSCH_CONF_MAC_MAX_FRAME_RETRIES
-#else
-#define TSCH_MAC_MAX_FRAME_RETRIES 7
-#endif
-
-/*********** Callbacks *********/
-
-#if BUILD_WITH_ORCHESTRA
-
-#ifndef TSCH_CALLBACK_NEW_TIME_SOURCE
-#define TSCH_CALLBACK_NEW_TIME_SOURCE orchestra_callback_new_time_source
-#endif /* TSCH_CALLBACK_NEW_TIME_SOURCE */
-
-#ifndef TSCH_CALLBACK_PACKET_READY
-#define TSCH_CALLBACK_PACKET_READY orchestra_callback_packet_ready
-#endif /* TSCH_CALLBACK_PACKET_READY */
-
-#endif /* BUILD_WITH_ORCHESTRA */
-
-/* Called by TSCH when switching time source */
-#ifdef TSCH_CALLBACK_NEW_TIME_SOURCE
-struct tsch_neighbor;
-void TSCH_CALLBACK_NEW_TIME_SOURCE(const struct tsch_neighbor *old, const struct tsch_neighbor *new);
-#endif
-
-/* Called by TSCH every time a packet is ready to be added to the send queue */
-#ifdef TSCH_CALLBACK_PACKET_READY
-void TSCH_CALLBACK_PACKET_READY(void);
-#endif
-
-/************ Types ***********/
-
-/* TSCH packet information */
-struct tsch_packet {
-  struct queuebuf *qb;  /* pointer to the queuebuf to be sent */
-  mac_callback_t sent; /* callback for this packet */
-  void *ptr; /* MAC callback parameter */
-  uint8_t transmissions; /* #transmissions performed for this packet */
-  uint8_t max_transmissions; /* maximal number of Tx before dropping the packet */
-  uint8_t ret; /* status -- MAC return code */
-  uint8_t header_len; /* length of header and header IEs (needed for link-layer security) */
-  uint8_t tsch_sync_ie_offset; /* Offset within the frame used for quick update of EB ASN and join priority */
-};
-
-/* TSCH neighbor information */
-struct tsch_neighbor {
-  /* Neighbors are stored as a list: "next" must be the first field */
-  struct tsch_neighbor *next;
-  linkaddr_t addr; /* MAC address of the neighbor */
-  uint8_t is_broadcast; /* is this neighbor a virtual neighbor used for broadcast (of data packets or EBs) */
-  uint8_t is_time_source; /* is this neighbor a time source? */
-  uint8_t backoff_exponent; /* CSMA backoff exponent */
-  uint8_t backoff_window; /* CSMA backoff window (number of slots to skip) */
-  uint8_t last_backoff_window; /* Last CSMA backoff window */
-  uint8_t tx_links_count; /* How many links do we have to this neighbor? */
-  uint8_t dedicated_tx_links_count; /* How many dedicated links do we have to this neighbor? */
-  /* Array for the ringbuf. Contains pointers to packets.
-   * Its size must be a power of two to allow for atomic put */
-  struct tsch_packet *tx_array[TSCH_QUEUE_NUM_PER_NEIGHBOR];
-  /* Circular buffer of pointers to packet. */
-  struct ringbufindex tx_ringbuf;
-};
 
 /***** External Variables *****/
