+++ os/net/mac/tsch/tsch-schedule.h
@@ -41,93 +41,7 @@
 /********** Includes **********/
 
 #include "contiki.h"
-#include "lib/list.h"
-#include "net/mac/tsch/tsch-private.h"
-#include "net/mac/tsch/tsch-queue.h"
-#include "net/mac/tsch/tsch-slot-operation.h"
 #include "net/linkaddr.h"
-
-/******** Configuration *******/
-
-/* Initializes TSCH with a 6TiSCH minimal schedule */
-#ifdef TSCH_SCHEDULE_CONF_WITH_6TISCH_MINIMAL
-#define TSCH_SCHEDULE_WITH_6TISCH_MINIMAL TSCH_SCHEDULE_CONF_WITH_6TISCH_MINIMAL
-#else
-#define TSCH_SCHEDULE_WITH_6TISCH_MINIMAL (!(BUILD_WITH_ORCHESTRA))
-#endif
-
-/* 6TiSCH Minimal schedule slotframe length */
-#ifdef TSCH_SCHEDULE_CONF_DEFAULT_LENGTH
-#define TSCH_SCHEDULE_DEFAULT_LENGTH TSCH_SCHEDULE_CONF_DEFAULT_LENGTH
-#else
-#define TSCH_SCHEDULE_DEFAULT_LENGTH 7
-#endif
-
-/* Max number of TSCH slotframes */
-#ifdef TSCH_SCHEDULE_CONF_MAX_SLOTFRAMES
-#define TSCH_SCHEDULE_MAX_SLOTFRAMES TSCH_SCHEDULE_CONF_MAX_SLOTFRAMES
-#else
-#define TSCH_SCHEDULE_MAX_SLOTFRAMES 4
-#endif
-
-/* Max number of links */
-#ifdef TSCH_SCHEDULE_CONF_MAX_LINKS
-#define TSCH_SCHEDULE_MAX_LINKS TSCH_SCHEDULE_CONF_MAX_LINKS
-#else
-#define TSCH_SCHEDULE_MAX_LINKS 32
-#endif
-
-/********** Constants *********/
-
-/* Link options */
-#define LINK_OPTION_TX              1
-#define LINK_OPTION_RX              2
-#define LINK_OPTION_SHARED          4
-#define LINK_OPTION_TIME_KEEPING    8
-
-/************ Types ***********/
-
-/* 802.15.4e link types.
- * LINK_TYPE_ADVERTISING_ONLY is an extra one: for EB-only links. */
-enum link_type { LINK_TYPE_NORMAL, LINK_TYPE_ADVERTISING, LINK_TYPE_ADVERTISING_ONLY };
-
-struct tsch_link {
-  /* Links are stored as a list: "next" must be the first field */
-  struct tsch_link *next;
-  /* Unique identifier */
-  uint16_t handle;
-  /* MAC address of neighbor */
-  linkaddr_t addr;
-  /* Slotframe identifier */
-  uint16_t slotframe_handle;
-  /* Identifier of Slotframe to which this link belongs
-   * Unused. */
-  /* uint8_t handle; */
-  /* Timeslot for this link */
-  uint16_t timeslot;
-  /* Channel offset for this link */
-  uint16_t channel_offset;
-  /* A bit string that defines
-   * b0 = Transmit, b1 = Receive, b2 = Shared, b3 = Timekeeping, b4 = reserved */
-  uint8_t link_options;
-  /* Type of link. NORMAL = 0. ADVERTISING = 1, and indicates
-     the link may be used to send an Enhanced beacon. */
-  enum link_type link_type;
-  /* Any other data for upper layers */
-  void *data;
-};
-
-struct tsch_slotframe {
-  /* Slotframes are stored as a list: "next" must be the first field */
-  struct tsch_slotframe *next;
-  /* Unique identifier */
-  uint16_t handle;
-  /* Number of timeslots in the slotframe.
-   * Stored as struct asn_divisor_t because we often need ASN%size */
-  struct tsch_asn_divisor_t size;
-  /* List of links belonging to this slotframe */
-  LIST_STRUCT(links_list);
-};
 
 /********** Functions *********/
