+++ os/net/mac/tsch/tsch.h
@@ -46,120 +46,40 @@ frequency hopping for enhanced reliability.
 
 #include "contiki.h"
 #include "net/mac/mac.h"
+#include "net/linkaddr.h"
+
+#include "net/mac/tsch/tsch-conf.h"
+#include "net/mac/tsch/tsch-const.h"
+#include "net/mac/tsch/tsch-types.h"
+#include "net/mac/tsch/tsch-adaptive-timesync.h"
+#include "net/mac/tsch/tsch-slot-operation.h"
+#include "net/mac/tsch/tsch-queue.h"
+#include "net/mac/tsch/tsch-log.h"
+#include "net/mac/tsch/tsch-packet.h"
 #include "net/mac/tsch/tsch-security.h"
+#include "net/mac/tsch/tsch-schedule.h"
+#if UIP_CONF_IPV6_RPL
+#include "net/mac/tsch/tsch-rpl.h"
+#endif /* UIP_CONF_IPV6_RPL */
 
-/******** Configuration *******/
+#if CONTIKI_TARGET_COOJA
+#include "lib/simEnvChange.h"
+#include "sys/cooja_mt.h"
+#endif /* CONTIKI_TARGET_COOJA */
 
-/* Max time before sending a unicast keep-alive message to the time source */
-#ifdef TSCH_CONF_KEEPALIVE_TIMEOUT
-#define TSCH_KEEPALIVE_TIMEOUT TSCH_CONF_KEEPALIVE_TIMEOUT
+/*********** Macros *********/
+
+/* Wait for a condition with timeout t0+offset. */
+#if CONTIKI_TARGET_COOJA
+#define BUSYWAIT_UNTIL_ABS(cond, t0, offset) \
+  while(!(cond) && RTIMER_CLOCK_LT(RTIMER_NOW(), (t0) + (offset))) { \
+    simProcessRunValue = 1; \
+    cooja_mt_yield(); \
+  };
 #else
-/* Time to desynch assuming a drift of 40 PPM (80 PPM between two nodes) and guard time of +/-1ms: 12.5s. */
-#define TSCH_KEEPALIVE_TIMEOUT (12 * CLOCK_SECOND)
-#endif
-
-/* With TSCH_ADAPTIVE_TIMESYNC enabled: keep-alive timeout used after reaching
- * accurate drift compensation. */
-#ifdef TSCH_CONF_MAX_KEEPALIVE_TIMEOUT
-#define TSCH_MAX_KEEPALIVE_TIMEOUT TSCH_CONF_MAX_KEEPALIVE_TIMEOUT
-#else
-#define TSCH_MAX_KEEPALIVE_TIMEOUT (60 * CLOCK_SECOND)
-#endif
-
-/* Max time without synchronization before leaving the PAN */
-#ifdef TSCH_CONF_DESYNC_THRESHOLD
-#define TSCH_DESYNC_THRESHOLD TSCH_CONF_DESYNC_THRESHOLD
-#else
-#define TSCH_DESYNC_THRESHOLD (2 * TSCH_MAX_KEEPALIVE_TIMEOUT)
-#endif
-
-/* Period between two consecutive EBs */
-#ifdef TSCH_CONF_EB_PERIOD
-#define TSCH_EB_PERIOD TSCH_CONF_EB_PERIOD
-#else
-#define TSCH_EB_PERIOD (16 * CLOCK_SECOND)
-#endif
-
-/* Max Period between two consecutive EBs */
-#ifdef TSCH_CONF_MAX_EB_PERIOD
-#define TSCH_MAX_EB_PERIOD TSCH_CONF_MAX_EB_PERIOD
-#else
-#define TSCH_MAX_EB_PERIOD (16 * CLOCK_SECOND)
-#endif
-
-/* Max acceptable join priority */
-#ifdef TSCH_CONF_MAX_JOIN_PRIORITY
-#define TSCH_MAX_JOIN_PRIORITY TSCH_CONF_MAX_JOIN_PRIORITY
-#else
-#define TSCH_MAX_JOIN_PRIORITY 32
-#endif
-
-/* Start TSCH automatically after init? If not, the upper layers
- * must call NETSTACK_MAC.on() to start it. Useful when the
- * application needs to control when the nodes are to start
- * scanning or advertising.*/
-#ifdef TSCH_CONF_AUTOSTART
-#define TSCH_AUTOSTART TSCH_CONF_AUTOSTART
-#else
-#define TSCH_AUTOSTART 1
-#endif
-
-/* Join only secured networks? (discard EBs with security disabled) */
-#ifdef TSCH_CONF_JOIN_SECURED_ONLY
-#define TSCH_JOIN_SECURED_ONLY TSCH_CONF_JOIN_SECURED_ONLY
-#else
-/* By default, set if LLSEC802154_ENABLED is also non-zero */
-#define TSCH_JOIN_SECURED_ONLY LLSEC802154_ENABLED
-#endif
-
-/* By default, join any PAN ID. Otherwise, wait for an EB from IEEE802154_PANID */
-#ifdef TSCH_CONF_JOIN_MY_PANID_ONLY
-#define TSCH_JOIN_MY_PANID_ONLY TSCH_CONF_JOIN_MY_PANID_ONLY
-#else
-#define TSCH_JOIN_MY_PANID_ONLY 1
-#endif
-
-/* The radio polling frequency (in Hz) during association process */
-#ifdef TSCH_CONF_ASSOCIATION_POLL_FREQUENCY
-#define TSCH_ASSOCIATION_POLL_FREQUENCY TSCH_CONF_ASSOCIATION_POLL_FREQUENCY
-#else
-#define TSCH_ASSOCIATION_POLL_FREQUENCY 100
-#endif
-
-/* When associating, check ASN against our own uptime (time in minutes)..
- * Useful to force joining only with nodes started roughly at the same time.
- * Set to the max number of minutes acceptable. */
-#ifdef TSCH_CONF_CHECK_TIME_AT_ASSOCIATION
-#define TSCH_CHECK_TIME_AT_ASSOCIATION TSCH_CONF_CHECK_TIME_AT_ASSOCIATION
-#else
-#define TSCH_CHECK_TIME_AT_ASSOCIATION 0
-#endif
-
-/* By default: initialize schedule from EB when associating, using the
- * slotframe and links Information Element */
-#ifdef TSCH_CONF_INIT_SCHEDULE_FROM_EB
-#define TSCH_INIT_SCHEDULE_FROM_EB TSCH_CONF_INIT_SCHEDULE_FROM_EB
-#else
-#define TSCH_INIT_SCHEDULE_FROM_EB 1
-#endif
-
-/* An ad-hoc mechanism to have TSCH select its time source without the
- * help of an upper-layer, simply by collecting statistics on received
- * EBs and their join priority. Disabled by default as we recomment
- * mapping the time source on the RPL preferred parent
- * (via tsch_rpl_callback_parent_switch) */
-#ifdef TSCH_CONF_AUTOSELECT_TIME_SOURCE
-#define TSCH_AUTOSELECT_TIME_SOURCE TSCH_CONF_AUTOSELECT_TIME_SOURCE
-#else
-#define TSCH_AUTOSELECT_TIME_SOURCE 0
-#endif /* TSCH_CONF_EB_AUTOSELECT */
-
-/* To include Sixtop Implementation */
-#ifdef TSCH_CONF_WITH_SIXTOP
-#define TSCH_WITH_SIXTOP TSCH_CONF_WITH_SIXTOP
-#else
-#define TSCH_WITH_SIXTOP 0
-#endif /* TSCH_CONF_EB_AUTOSELECT */
+#define BUSYWAIT_UNTIL_ABS(cond, t0, offset) \
+  while(!(cond) && RTIMER_CLOCK_LT(RTIMER_NOW(), (t0) + (offset))) ;
+#endif /* CONTIKI_TARGET_COOJA */
 
 /*********** Callbacks *********/
 
@@ -180,6 +100,18 @@ frequency hopping for enhanced reliability.
 
 #endif /* UIP_CONF_IPV6_RPL */
 
+#if BUILD_WITH_ORCHESTRA
+
+#ifndef TSCH_CALLBACK_NEW_TIME_SOURCE
+#define TSCH_CALLBACK_NEW_TIME_SOURCE orchestra_callback_new_time_source
+#endif /* TSCH_CALLBACK_NEW_TIME_SOURCE */
+
+#ifndef TSCH_CALLBACK_PACKET_READY
+#define TSCH_CALLBACK_PACKET_READY orchestra_callback_packet_ready
+#endif /* TSCH_CALLBACK_PACKET_READY */
+
+#endif /* BUILD_WITH_ORCHESTRA */
+
 /* Called by TSCH when joining a network */
 #ifdef TSCH_CALLBACK_JOINING_NETWORK
 void TSCH_CALLBACK_JOINING_NETWORK();
@@ -188,6 +120,28 @@ void TSCH_CALLBACK_JOINING_NETWORK();
 /* Called by TSCH when leaving a network */
 #ifdef TSCH_CALLBACK_LEAVING_NETWORK
 void TSCH_CALLBACK_LEAVING_NETWORK();
+#endif
+
+/* Called by TSCH after sending a keep-alive */
+#ifdef TSCH_CALLBACK_KA_SENT
+void TSCH_CALLBACK_KA_SENT();
+#endif
+
+/* Called by TSCH form interrupt after receiving a frame, enabled upper-layer to decide
+ * whether to ACK or NACK */
+#ifdef TSCH_CALLBACK_DO_NACK
+int TSCH_CALLBACK_DO_NACK(struct tsch_link *link, linkaddr_t *src, linkaddr_t *dst);
+#endif
+
+/* Called by TSCH when switching time source */
+#ifdef TSCH_CALLBACK_NEW_TIME_SOURCE
+struct tsch_neighbor;
+void TSCH_CALLBACK_NEW_TIME_SOURCE(const struct tsch_neighbor *old, const struct tsch_neighbor *new);
+#endif
+
+/* Called by TSCH every time a packet is ready to be added to the send queue */
+#ifdef TSCH_CALLBACK_PACKET_READY
+void TSCH_CALLBACK_PACKET_READY(void);
 #endif
 
 /***** External Variables *****/
@@ -200,6 +154,29 @@ extern int tsch_is_associated;
 extern int tsch_is_pan_secured;
 /* The TSCH MAC driver */
 extern const struct mac_driver tschmac_driver;
+/* 802.15.4 broadcast MAC address */
+extern const linkaddr_t tsch_broadcast_address;
+/* The address we use to identify EB queue */
+extern const linkaddr_t tsch_eb_address;
+/* The current Absolute Slot Number (ASN) */
+extern struct tsch_asn_t tsch_current_asn;
+extern uint8_t tsch_join_priority;
+extern struct tsch_link *current_link;
+/* TSCH channel hopping sequence */
+extern uint8_t tsch_hopping_sequence[TSCH_HOPPING_SEQUENCE_MAX_LEN];
+extern struct tsch_asn_divisor_t tsch_hopping_sequence_length;
+/* TSCH timeslot timing (in rtimer ticks) */
+extern rtimer_clock_t tsch_timing[tsch_ts_elements_count];
+/* Statistics on the current session */
+extern unsigned long tx_count;
+extern unsigned long rx_count;
+extern unsigned long sync_count;
+
+/* TSCH processes */
+PROCESS_NAME(tsch_process);
+PROCESS_NAME(tsch_send_eb_process);
+PROCESS_NAME(tsch_pending_events_process);
+
 
 /********** Functions *********/
 
@@ -213,6 +190,12 @@ void tsch_set_ka_timeout(uint32_t timeout);
 void tsch_set_coordinator(int enable);
 /* Set the pan as secured or not */
 void tsch_set_pan_secured(int enable);
+/* Set TSCH to send a keepalive message after TSCH_KEEPALIVE_TIMEOUT */
+void tsch_schedule_keepalive(void);
+/* Set TSCH to send a keepalive message immediately */
+void tsch_schedule_keepalive_immediately(void);
+/* Leave the TSCH network */
+void tsch_disassociate(void);
 
 #endif /* __TSCH_H__ */
 /** @} */
